feat(k8s): add worker-deployment.yaml file to define a deployment for the worker component in Kubernetes
The worker-deployment.yaml file is added to the k8s directory. It defines a Deployment resource for the worker component in Kubernetes. The Deployment has a selector that matches labels with the component set to "worker". The template section specifies the pod template for the Deployment, with labels set to "worker". The pod template includes a container named "worker" with the image "stephengrider/multi-worker".